Introduce cache for local metadata availability
This commit adds a cache for local metadata availability. Since we introduced the heuristic to determine if metadata
is available locally, we could hit the persistent store multiple times for the same module during a build, which
introduced a performance regression. This commit adds the ability to cache the answer of availability for a build,
making it possible to do a short check before returning.